pyproject.toml
[tool.poetry]
name = "fastapi-layered-architecture"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pre-commit = "^4.0.1"
sqlalchemy = "^2.0.36"
poetry-dotenv = "^0.4.0"
python-dotenv = "^1.0.1"
dependency-injector = "^4.43.0"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
